GH-281: New filter for $feed [Master]

+++ admin/class-fastcgi-purger.php
@@ -57,7 +57,7 @@ class FastCGI_Purger extends Purger {
 
 				$this->delete_cache_file_for( $_url_purge );
 
-				if ( $feed ) {
+				if ( $feed && ! empty( $nginx_helper_admin->options['purge_feeds'] ) ) {
 
 					$feed_url = rtrim( $_url_purge_base, '/' ) . '/feed/';
 					$this->delete_cache_file_for( $feed_url );
@@ -79,7 +79,7 @@ class FastCGI_Purger extends Purger {
 
 				$this->do_remote_get( $_url_purge );
 
-				if ( $feed ) {
+				if ( $feed && ! empty( $nginx_helper_admin->options['purge_feeds'] ) ) {
 
 					$feed_url = rtrim( $_url_purge_base, '/' ) . '/feed/';
 					$this->do_remote_get( $feed_url );
